Refactor benchmark - put test library inside library
