+++ library/Java/Lang/String/StringBenchmark.cpp
@@ -24,10 +24,9 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "../../../../benchmark/benchmark.hpp"
+#include "../../../../misc/benchmark/benchmark.hpp"
 
 #include "String.hpp"
-#include <string>
 
 using namespace Java::Lang;
