+++ library/Java/Util/HashMap/HashMapBenchmark.cpp
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#include "../../../../benchmark/benchmark.hpp"
+#include "../../../../misc/benchmark/benchmark.hpp"
 #include "../HashMap/HashMap.hpp"
 #include <unordered_map>
